+++ src/features/screens/mapView.screen.js
@@ -18,6 +18,7 @@ import styled from 'styled-components/native';
 import { Spacer } from '../../components/spacers/spacer.component';
 import SearchInput from '../../features/searchBar/SeachInput';
 
+
 export default function MapScreen() {
     //state for user location
     const [location, setLocation] = useState(null);
@@ -30,6 +31,7 @@ export default function MapScreen() {
     //state for search query
     const [searchResults, setSearchResults] = useState([]);
     //need a state for after the search is completed
+
 
     const radius = 5 * 1609;
     const prevRestaurantsRef = useRef([]);
@@ -153,6 +155,7 @@ export default function MapScreen() {
         //LIST VIEW STATE
         return (
             <SafeArea>
+
                 <SearchContainer>
                     {/* pass the function as props */}
                     <SearchInput
@@ -302,3 +305,4 @@ const styles = StyleSheet.create({
         elevation: 3,
     },
 });
+
